requirements.txt
requests
python-dotenv
pytest
flake8
black
